Remove binding from ConfigFileApplicationListener
Closes gh-8101
